# perf(dracut-functions.sh): optimise get_persistent_dev()