99base install() goes from like
```
real    0m0.350s
user    0m0.250s
sys     0m0.119s
```
to like
```
real    0m0.216s
user    0m0.162s
sys     0m0.076s
```
on my system in the average case.

Microbenchmarks show, this is appx. an easy-average case:
```
$ hyperfine -S'bash --norc' 'gget_persistent_dev /dev/zd0' 'mget_persistent_dev /dev/zd0' 'old_get_persistent_dev /dev/zd0'
Benchmark 1: gget_persistent_dev /dev/zd0
  Time (mean ± σ):       7.2 ms ±   0.3 ms    [User: 6.7 ms, System: 4.0 ms]
  Range (min … max):     5.3 ms …   8.4 ms    307 runs

Benchmark 2: mget_persistent_dev /dev/zd0
  Time (mean ± σ):       7.1 ms ±   0.6 ms    [User: 6.2 ms, System: 3.1 ms]
  Range (min … max):     4.1 ms …   7.9 ms    301 runs

Benchmark 3: old_get_persistent_dev /dev/zd0
  Time (mean ± σ):      22.3 ms ±   1.4 ms    [User: 19.1 ms, System: 5.3 ms]
  Range (min … max):    19.7 ms …  27.3 ms    127 runs

Summary
  'mget_persistent_dev /dev/zd0' ran
    1.01 ± 0.10 times faster than 'gget_persistent_dev /dev/zd0'
    3.14 ± 0.34 times faster than 'old_get_persistent_dev /dev/zd0'
```

For best case (this is the first device in the glob):
```
$ hyperfine -S'bash --norc' 'gget_persistent_dev /dev/zd272p1' 'mget_persistent_dev /dev/zd272p1' 'old_get_persistent_dev /dev/zd272p1'
+ hyperfine --style=basic '-Sbash --norc' 'gget_persistent_dev /dev/zd272p1' 'mget_persistent_dev /dev/zd272p1' 'old_get_persistent_dev /dev/zd272p1'
Benchmark 1: gget_persistent_dev /dev/zd272p1
  Time (mean ± σ):       6.5 ms ±   0.8 ms    [User: 6.1 ms, System: 3.8 ms]
  Range (min … max):     4.4 ms …   7.9 ms    284 runs

Benchmark 2: mget_persistent_dev /dev/zd272p1
  Time (mean ± σ):       6.9 ms ±   0.5 ms    [User: 5.8 ms, System: 3.2 ms]
  Range (min … max):     4.5 ms …   8.1 ms    299 runs

Benchmark 3: old_get_persistent_dev /dev/zd272p1
  Time (mean ± σ):       8.4 ms ±   0.7 ms    [User: 7.0 ms, System: 2.1 ms]
  Range (min … max):     6.1 ms …   9.6 ms    246 runs

Summary
  'gget_persistent_dev /dev/zd272p1' ran
    1.06 ± 0.15 times faster than 'mget_persistent_dev /dev/zd272p1'
    1.28 ± 0.19 times faster than 'old_get_persistent_dev /dev/zd272p1'
```

And for worst case (`/dev/zero` doesn't have a corresponding device blockdev in the glob):
```
$ hyperfine -S'bash --norc' 'gget_persistent_dev /dev/zero' 'mget_persistent_dev /dev/zero' 'old_get_persistent_dev /dev/zero'
Benchmark 1: gget_persistent_dev /dev/zero
  Time (mean ± σ):       7.0 ms ±   0.7 ms    [User: 6.5 ms, System: 4.0 ms]
  Range (min … max):     5.0 ms …   8.5 ms    269 runs

Benchmark 2: mget_persistent_dev /dev/zero
  Time (mean ± σ):       6.7 ms ±   0.9 ms    [User: 6.0 ms, System: 2.8 ms]
  Range (min … max):     4.8 ms …   8.8 ms    275 runs

Benchmark 3: old_get_persistent_dev /dev/zero
  Time (mean ± σ):     791.9 ms ±  44.6 ms    [User: 646.9 ms, System: 215.4 ms]
  Range (min … max):   723.1 ms … 842.1 ms    10 runs

Summary
  'mget_persistent_dev /dev/zero' ran
    1.04 ± 0.17 times faster than 'gget_persistent_dev /dev/zero'
  117.86 ± 17.18 times faster than 'old_get_persistent_dev /dev/zero'
```

Or, in short:
  * 20 -> 6 ms
  * 8 -> 6 ms
  * 800 -> 7 ms

In the cases that are run in the 99base setup, I saw baseline speeds of 40-50ms, and it's run like 3 times there for me. If you have a less weird (non-root-on-ZFS) or more weird (wide btrfs) setup you'll likely see a better improvement.

`get_persistent_dev()` is only used in `module-setup.sh`s

## Checklist
- [x] I have tested it locally
- [x] I have reviewed and updated any documentation if relevant
- [x] I am providing new code and test(s) for it