@nabijaczleweli 2022-10-22
Admittedly, it only happened when I extended the stat cmdline to 175x its original size on my mid-sized system (lsblk is 31 devices with grep -v p and 92 without), but I had been worried about just blowing the stat environment block size.

The contenders are:
```sh
sget_persistent_dev() {
    local _dev _pol
    _dev=$(stat -L -c $'%t:%T' "$1" 2> /dev/null)
    [ -z "$_dev" ] && return
    if [[ -n $persistent_policy ]]; then
        _pol="/dev/disk/${persistent_policy}/*"
    else
        _pol=
    fi
    # shellcheck disable=SC2086
    stat -L -c $'%n\t%t:%T' \
        $_pol \
        /dev/mapper/* \
        /dev/disk/by-uuid/* \
        /dev/disk/by-label/* \
        /dev/disk/by-partuuid/* \
        /dev/disk/by-partlabel/* \
        /dev/disk/by-id/* \
        /dev/disk/by-path/* 2> /dev/null \
        | awk -v_dev="$_dev" -vdfl="$1" '$0 ~ ("\t" _dev "$") {--NF; print; dfl=0; exit}  END {if(dfl) print dfl}'
}

pget_persistent_dev() {
    local _dev _pol
    _dev=$(stat -L -c $'%t:%T' "$1" 2> /dev/null)
    [ -z "$_dev" ] && return
    if [[ -n $persistent_policy ]]; then
        _pol="/dev/disk/${persistent_policy}/*"
    else
        _pol=
    fi
    # shellcheck disable=SC2086
    printf '%s\0' \
        $_pol \
        /dev/mapper/* \
        /dev/disk/by-uuid/* \
        /dev/disk/by-label/* \
        /dev/disk/by-partuuid/* \
        /dev/disk/by-partlabel/* \
        /dev/disk/by-id/* \
        /dev/disk/by-path/* \
        | xargs -0 stat -L -c $'%n\t%t:%T' 2> /dev/null \
        | awk -v_dev="$_dev" -vdfl="$1" '$0 ~ ("\t" _dev "$") {--NF; print; dfl=0; exit}  END {if(dfl) print dfl}'
}

fget_persistent_dev() {
    local _dev _pol
    _dev=$(stat -L -c $'%t:%T' "$1" 2> /dev/null)
    [ -z "$_dev" ] && return
    if [[ -n $persistent_policy ]]; then
        _pol="/dev/disk/${persistent_policy}"
    else
        _pol=
    fi
    # shellcheck disable=SC2086
    find \
        $_pol \
        /dev/mapper \
        /dev/disk/by-uuid \
        /dev/disk/by-label \
        /dev/disk/by-partuuid \
        /dev/disk/by-partlabel \
        /dev/disk/by-id \
        /dev/disk/by-path \
        -print0  \
        | xargs -0 stat -L -c $'%n\t%t:%T' 2> /dev/null \
        | awk -v_dev="$_dev" -vdfl="$1" '$0 ~ ("\t" _dev "$") {--NF; print; dfl=0; exit}  END {if(dfl) print dfl}'
}
```

And the results:
```
$ hyperfine -S'bash --norc' \
> 'sget_persistent_dev /dev/zd0' \
> 'pget_persistent_dev /dev/zd0' \
> 'fget_persistent_dev /dev/zd0'
Benchmark 1: sget_persistent_dev /dev/zd0
  Time (mean ± σ):       6.9 ms ±   0.7 ms    [User: 5.9 ms, System: 3.0 ms]
  Range (min … max):     4.5 ms …   7.9 ms    289 runs

Benchmark 2: pget_persistent_dev /dev/zd0
  Time (mean ± σ):       7.8 ms ±   0.8 ms    [User: 8.1 ms, System: 3.3 ms]
  Range (min … max):     5.4 ms …   9.8 ms    276 runs

Benchmark 3: fget_persistent_dev /dev/zd0
  Time (mean ± σ):       7.3 ms ±   1.0 ms    [User: 7.1 ms, System: 3.7 ms]
  Range (min … max):     5.1 ms …   8.9 ms    306 runs

Summary
  'sget_persistent_dev /dev/zd0' ran
    1.06 ± 0.18 times faster than 'fget_persistent_dev /dev/zd0'
    1.14 ± 0.17 times faster than 'pget_persistent_dev /dev/zd0'
```

```
$ hyperfine -S'bash --norc' \
> 'sget_persistent_dev /dev/zd272p1' \
> 'pget_persistent_dev /dev/zd272p1' \
> 'fget_persistent_dev /dev/zd272p1'
Benchmark 1: sget_persistent_dev /dev/zd272p1
  Time (mean ± σ):       6.7 ms ±   1.1 ms    [User: 5.8 ms, System: 3.0 ms]
  Range (min … max):     4.3 ms …   8.1 ms    289 runs

Benchmark 2: pget_persistent_dev /dev/zd272p1
  Time (mean ± σ):       6.7 ms ±   0.8 ms    [User: 7.6 ms, System: 2.4 ms]
  Range (min … max):     5.1 ms …   8.4 ms    373 runs

Benchmark 3: fget_persistent_dev /dev/zd272p1
  Time (mean ± σ):       6.7 ms ±   1.0 ms    [User: 6.5 ms, System: 3.7 ms]
  Range (min … max):     5.1 ms …   8.9 ms    296 runs

Summary
  'sget_persistent_dev /dev/zd272p1' ran
    1.00 ± 0.20 times faster than 'pget_persistent_dev /dev/zd272p1'
    1.00 ± 0.22 times faster than 'fget_persistent_dev /dev/zd272p1'
```

```
$ hyperfine -S'bash --norc' \
> 'sget_persistent_dev /dev/zero' \
> 'pget_persistent_dev /dev/zero' \
> 'fget_persistent_dev /dev/zero'
Benchmark 1: sget_persistent_dev /dev/zero
  Time (mean ± σ):       7.3 ms ±   0.8 ms    [User: 6.1 ms, System: 3.2 ms]
  Range (min … max):     5.4 ms …   8.7 ms    374 runs

Benchmark 2: pget_persistent_dev /dev/zero
  Time (mean ± σ):       7.0 ms ±   0.7 ms    [User: 7.4 ms, System: 2.8 ms]
  Range (min … max):     5.9 ms …   9.2 ms    344 runs

Benchmark 3: fget_persistent_dev /dev/zero
  Time (mean ± σ):       7.9 ms ±   0.9 ms    [User: 7.6 ms, System: 3.7 ms]
  Range (min … max):     5.8 ms …  10.3 ms    329 runs

Summary
  'pget_persistent_dev /dev/zero' ran
    1.04 ± 0.15 times faster than 'sget_persistent_dev /dev/zero'
    1.12 ± 0.17 times faster than 'fget_persistent_dev /dev/zero'
```

Which I'd classify as "who the hell knows" because, really, who does, it's all roughly within jitter, so I've rebased and updated to use printf because I like it more and it avoids an exec.