@nabijaczleweli 2022-12-03
world's stinkiest bot 2022 (real)